Move tests to corresponding namespaces
