+++ Assets/Tests/EditMode/Infrastructure/Factories/Components/ComponentFactoryTests.cs
@@ -2,9 +2,10 @@ using System;
 using FluentAssertions;
 using NUnit.Framework;
 using UnityEngine;
-using WizardSpells.Infrastructure.Factories.Component;
+using WizardSpells.Infrastructure.Factories.Components;
+using WizardSpells.Tests.EditMode.TestUtilities;
 
-namespace WizardSpells.Tests.EditMode
+namespace WizardSpells.Tests.EditMode.Infrastructure.Factories.Components
 {
     public class ComponentFactoryTests
     {
