+++ Assets/Tests/EditMode/TestUtilities/Create.cs
@@ -2,14 +2,14 @@ using System;
 using System.Linq;
 using UnityEngine;
 using WizardSpells.Data.Scene;
-using WizardSpells.Infrastructure.Factories.Component;
+using WizardSpells.Infrastructure.Factories.Components;
 using WizardSpells.Infrastructure.GameStates.InitialSettingStrategies;
 using WizardSpells.Services.SceneManagement;
 using WizardSpells.Utilities.Patterns.State.Machines;
 using WizardSpells.Utilities.Patterns.Strategy;
 using Zenject;
 
-namespace WizardSpells.Tests.EditMode
+namespace WizardSpells.Tests.EditMode.TestUtilities
 {
     public static class Create
     {
@@ -19,7 +19,7 @@ namespace WizardSpells.Tests.EditMode
         public static SceneStrategyProvider SceneStrategyProvider(IStrategy defaultStrategy = default, 
             params (SceneName SceneName, IStrategy Instance)[] strategies) =>
             new(strategies?.ToDictionary(strategy => strategy.SceneName, strategy => strategy.Instance)
-                    .SubstituteIfDefault(), defaultStrategy.SubstituteIfDefault());
+                .SubstituteIfDefault(), defaultStrategy.SubstituteIfDefault());
 
         public static GameObject GameObject(params Type[] componentTypesToAttach)
         {
@@ -29,8 +29,8 @@ namespace WizardSpells.Tests.EditMode
             return newGameObject;
         }
 
-        public static ComponentFactory<TComponent> ComponentFactory<TComponent>(GameObject originalGameObject = default, 
-            Transform objectsParent = default)  where TComponent : Component =>
+        public static ComponentFactory<TComponent> ComponentFactory<TComponent>(GameObject originalGameObject = default,
+            Transform objectsParent = default) where TComponent : Component =>
             new(Substitute.FactoryConfig(originalGameObject.NewIfDefault()), objectsParent.SubstituteIfDefault());
 
         public static DependentComponentFactory<TComponent> DependentComponentFactory<TComponent>(
@@ -39,6 +39,6 @@ namespace WizardSpells.Tests.EditMode
             new(Substitute.FactoryConfig(originalGameObject.NewIfDefault()), objectsParent.SubstituteIfDefault(),
                 container.SubstituteIfDefault());
 
-        public static T NewIfDefault<T>(this T obj) where T : class, new() => obj ?? new();
+        public static T NewIfDefault<T>(this T obj) where T : class, new() => obj ?? new T();
     }
 }
