+++ Assets/Tests/PlayMode/TestUtilities/Create.cs
@@ -1,0 +1,10 @@
+﻿using NSubstitute;
+using WizardSpells.Services.SceneManagement;
+
+namespace WizardSpells.Tests.PlayMode.TestUtilities
+{
+    public static class Create
+    {
+        public static SceneLoader SceneLoader() => Substitute.For<SceneLoader>();
+    }
+}
